fix bug on failed connection
Sockets on outgoing connections were not being destroyed properly when failing to connect. This is because `peer.socket` would not be initilaized until after a successful connection, whereas `peer.close()` would only destroy `this.socket`. The fix assigns a value to `peer.socket` during connection attempts, so that the socket will be properly destroyed on error.