docs: restructure readme, link to wiki
This commit restructures the README to provide a summary of what xud is
and how to install and get started with it, while moving more
detailed instructions to the wiki and linking to it. This aims to
more closely follow GitHub recommendations for README.md and using the
wiki. Going forward, updating project documentation can be done without
creating commits or pull requests.

https://guides.github.com/features/wikis/